Flatten validation error list and reduce complexity
